requirements.txt
scipy
numpy
matplotlib==3.1.2
opencv_python
torch
torchvision
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0
labelme==3.16.7
albumentations
selenium